oops thats awkward

+++ ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: cyildiri <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/09/24 17:24:57 by cyildiri          #+#    #+#             */
-/*   Updated: 2016/09/26 12:26:30 by cyildiri         ###   ########.fr       */
+/*   Updated: 2016/10/17 20:27:31 by cyildiri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
 	int count;
